Add instrument arrayList to shop

+++ tunes_r_us/src/main/java/interfaces/ISell.java
@@ -2,5 +2,5 @@ package interfaces;
 
 public interface ISell {
 
-    int calculateMarkUp();
+    double calculateMarkUp();
 }
